Move business details code to modules
