+++ src/client/modules/Companies/CollectionList/labels.js
@@ -112,25 +112,6 @@ const exportPotentialLabels = {
   },
 }
 
-const exportSegmentsLabels = {
-  hep: 'High export potential',
-  'non-hep': 'Not high export potential',
-}
-const exportSubSegmentsLabels = {
-  sustain_nurture_and_grow: 'Sustain: nurture & grow',
-  sustain_develop_export_capability: 'Sustain: develop export capability',
-  sustain_communicate_benefits: 'Sustain: communicate benefits',
-  sustain_increase_competitiveness: 'Sustain: increase competitiveness',
-  reassure_nurture_and_grow: 'Reassure: nurture & grow',
-  reassure_develop_export_capability: 'Reassure: develop export capability',
-  reassure_leave_be: 'Reassure: leave be',
-  reassure_change_the_game: 'Reassure: change the game',
-  promote_develop_export_capability: 'Promote: develop export capability',
-  promote_communicate_benefits: 'Promote: communicate benefits',
-  promote_change_the_game: 'Promote: change the game',
-  challenge: 'Challenge',
-}
-
 export default {
   companyDetailsLabels,
   companyTypeOptions,
@@ -142,6 +123,4 @@ export default {
   aboutLabels,
   businessHierarchyLabels,
   exportPotentialLabels,
-  exportSegmentsLabels,
-  exportSubSegmentsLabels,
 }
