+++ src/client/modules/Companies/CompanyBusinessDetails/SectionRegion.jsx
@@ -2,8 +2,8 @@ import React from 'react'
 import PropTypes from 'prop-types'
 import Link from '@govuk-react/link'
 
-import { SummaryTable } from '../../../../../client/components/'
-import urls from '../../../../../lib/urls'
+import { SummaryTable } from '../../../components'
+import urls from '../../../../lib/urls'
 
 const SectionRegion = ({ company, isArchived }) =>
   !!company.ukBased ? (
