+++ src/client/modules/Companies/CompanyBusinessDetails/labels.js
@@ -1,0 +1,19 @@
+export const exportSegmentsLabels = {
+  hep: 'High export potential',
+  'non-hep': 'Not high export potential',
+}
+
+export const exportSubSegmentsLabels = {
+  sustain_nurture_and_grow: 'Sustain: nurture & grow',
+  sustain_develop_export_capability: 'Sustain: develop export capability',
+  sustain_communicate_benefits: 'Sustain: communicate benefits',
+  sustain_increase_competitiveness: 'Sustain: increase competitiveness',
+  reassure_nurture_and_grow: 'Reassure: nurture & grow',
+  reassure_develop_export_capability: 'Reassure: develop export capability',
+  reassure_leave_be: 'Reassure: leave be',
+  reassure_change_the_game: 'Reassure: change the game',
+  promote_develop_export_capability: 'Promote: develop export capability',
+  promote_communicate_benefits: 'Promote: communicate benefits',
+  promote_change_the_game: 'Promote: change the game',
+  challenge: 'Challenge',
+}
